refactor and use lru for image cache
