+++ src/Layer/VSkia.hpp
@@ -15,12 +15,12 @@
  */
 #pragma once
 #include "Layer/AttrSerde.hpp"
-#include "Layer/Effects.hpp"
 #include "Layer/Core/VUtils.hpp"
 #include "Layer/Core/VType.hpp"
 #include "Layer/Core/Attrs.hpp"
 #include "Layer/Scene.hpp"
 #include "Layer/SkSL.hpp"
+
 #include "Math/Algebra.hpp"
 #include "Math/Math.hpp"
 #include "Utility/Log.hpp"
@@ -28,6 +28,7 @@
 
 #include <core/SkColorFilter.h>
 #include <core/SkRefCnt.h>
+#include <core/SkSamplingOptions.h>
 #include <core/SkScalar.h>
 #include <core/SkShader.h>
 #include <core/SkPathEffect.h>
@@ -50,8 +51,6 @@
 #include <unordered_map>
 #include <string>
 using namespace VGG;
-
-extern std::unordered_map<std::string, sk_sp<SkImage>> g_skiaImageRepo;
 
 inline SkRect toSkRect(const VGG::Bound& bound)
 {
@@ -161,410 +160,6 @@ inline SkPathOp toSkPathOp(VGG::EBoolOp blop)
   SWITCH_MAP_ITEM_DEF(VGG::BO_INTERSECTION, SkPathOp::kIntersect_SkPathOp)
   SWITCH_MAP_ITEM_DEF(VGG::BO_EXECLUSION, SkPathOp::kXOR_SkPathOp)
   SWITCH_MAP_ITEM_END(SkPathOp::kUnion_SkPathOp)
-}
-
-inline sk_sp<SkImage> loadImage(const std::string& imageGUID, const ResourceRepo& repo)
-{
-  sk_sp<SkImage> image;
-  if (imageGUID.empty())
-    return image;
-  std::string guid = imageGUID;
-  if (auto pos = guid.find("./"); pos != std::string::npos && pos == 0)
-  {
-    // remove current dir notation
-    guid = guid.substr(2);
-  }
-  if (auto it = g_skiaImageRepo.find(guid); it != g_skiaImageRepo.end())
-  {
-    image = it->second;
-  }
-  else
-  {
-    auto repo = Scene::getResRepo();
-    if (auto it = repo.find(guid); it != repo.end())
-    {
-      auto data = SkData::MakeWithCopy(it->second.data(), it->second.size());
-      if (!data)
-      {
-        WARN("Make SkData failed");
-        return image;
-      }
-      sk_sp<SkImage> skImage = SkImages::DeferredFromEncodedData(data);
-      if (!skImage)
-      {
-        WARN("Make SkImage failed.");
-        return image;
-      }
-      g_skiaImageRepo[guid] = skImage;
-      image = skImage;
-    }
-    else
-    {
-      WARN("Cannot find %s from resources repository", guid.c_str());
-    }
-  }
-  return image;
-}
-
-sk_sp<SkColorFilter> makeColorFilter(const ImageFilter& imageFilter);
-
-inline SkMatrix makeMatrix(
-  const Bound&                          bound,
-  const glm::vec2&                      f,
-  const glm::vec2&                      t,
-  const std::variant<float, glm::vec2>& ellipse)
-{
-  const auto from = bound.map(bound.size() * f);
-  const auto to = bound.map(bound.size() * t);
-
-  auto theta = [](const glm::vec2& from, const glm::vec2& to)
-  {
-    const auto d = to - from;
-    return std::atan2(d.y, d.x);
-  };
-
-  auto ratio = 1.f;
-  std::visit(
-    Overloaded{ [&](const float& f) { ratio = f; },
-                [&](const glm::vec2& p)
-                {
-                  auto       pp = bound.map(bound.size() * p);
-                  const auto a = glm::distance(from, to);
-                  const auto b = glm::distance(from, pp);
-                  ratio = (a == 0.f) ? 1.f : b / a;
-                } },
-    ellipse);
-  SkMatrix mat = SkMatrix::I();
-  mat.postTranslate(-from.x, -from.y);
-  mat.postScale(1.0, ratio);
-  mat.postRotate(rad2deg(theta(from, to)));
-  mat.postTranslate(from.x, from.y);
-  return mat;
-}
-
-inline sk_sp<SkShader> makeFitPattern(const Bound& bound, const PatternFit& p)
-{
-  auto img = loadImage(p.guid, Scene::getResRepo());
-  if (!img)
-    return nullptr;
-  SkImageInfo mi = img->imageInfo();
-  float       width = bound.width();
-  float       height = bound.height();
-  float       sx = (float)width / mi.width();
-  float       sy = (float)height / mi.height();
-  auto        m = glm::mat3{ 1.0 };
-  float       s = std::min(sx, sy);
-  if (sx < sy)
-  {
-    m = glm::translate(m, { 0, (height - s * mi.height()) / 2 });
-  }
-  else
-  {
-    m = glm::translate(m, { (width - s * mi.width()) / 2, 0 });
-  }
-  m = glm::scale(m, { s, s });
-  m = glm::translate(m, { mi.width() / 2, mi.height() / 2 });
-  m = glm::rotate(m, p.rotation);
-  m = glm::translate(m, { -mi.width() / 2, -mi.height() / 2 });
-  SkSamplingOptions opt(SkFilterMode::kLinear);
-  SkTileMode        modeX = SkTileMode::kDecal;
-  SkTileMode        modeY = SkTileMode::kDecal;
-  const auto        mat = toSkMatrix(m);
-  auto              shader = img->makeShader(modeX, modeY, opt, &mat);
-  if (auto colorFilter = makeColorFilter(p.imageFilter); shader && colorFilter)
-  {
-    return shader->makeWithColorFilter(colorFilter);
-  }
-  return shader;
-}
-inline sk_sp<SkShader> makeFillPattern(const Bound& bound, const PatternFill& p)
-{
-  auto img = loadImage(p.guid, Scene::getResRepo());
-
-  if (!img)
-    return nullptr;
-  SkImageInfo mi = img->imageInfo();
-  float       width = bound.width();
-  float       height = bound.height();
-  float       sx = (float)width / mi.width();
-  float       sy = (float)height / mi.height();
-  auto        m = glm::mat3{ 1.0 };
-  const float s = std::max(sx, sy);
-  if (sx > sy)
-  {
-    m = glm::translate(m, { 0, (height - s * mi.height()) / 2.f });
-  }
-  else
-  {
-    m = glm::translate(m, { (width - s * mi.width()) / 2.f, 0 });
-  }
-  m = glm::scale(m, { s, s });
-  m = glm::translate(m, { mi.width() / 2, mi.height() / 2 });
-  m = glm::rotate(m, p.rotation);
-  m = glm::translate(m, { -mi.width() / 2, -mi.height() / 2 });
-  SkSamplingOptions opt(SkFilterMode::kLinear);
-  SkTileMode        modeX = SkTileMode::kDecal;
-  SkTileMode        modeY = SkTileMode::kDecal;
-  const auto        mat = toSkMatrix(m);
-  auto              shader = img->makeShader(modeX, modeY, opt, &mat);
-  if (auto colorFilter = makeColorFilter(p.imageFilter); shader && colorFilter)
-  {
-    return shader->makeWithColorFilter(colorFilter);
-  }
-  return shader;
-}
-
-inline sk_sp<SkShader> makeStretchPattern(const Bound& bound, const PatternStretch& p)
-{
-  auto img = loadImage(p.guid, Scene::getResRepo());
-  if (!img)
-    return nullptr;
-  SkImageInfo mi = img->imageInfo();
-  float       width = bound.width();
-  float       height = bound.height();
-  auto        m = glm::mat3{ 1.0 };
-  m = glm::scale(m, { width, height });
-  m *= p.transform.matrix();
-  m = glm::scale(m, { 1.f / mi.width(), 1.f / mi.height() });
-  const auto        mat = toSkMatrix(m);
-  SkSamplingOptions opt(SkFilterMode::kLinear);
-  SkTileMode        modeX = SkTileMode::kDecal;
-  SkTileMode        modeY = SkTileMode::kDecal;
-  auto              shader = img->makeShader(modeX, modeY, opt, &mat);
-  if (auto colorFilter = makeColorFilter(p.imageFilter); shader && colorFilter)
-  {
-    return shader->makeWithColorFilter(colorFilter);
-  }
-  return shader;
-}
-
-inline sk_sp<SkShader> makeTilePattern(const Bound& bound, const PatternTile& p)
-{
-  auto img = loadImage(p.guid, Scene::getResRepo());
-
-  if (!img)
-    return nullptr;
-  SkTileMode modeX = SkTileMode::kDecal;
-  SkTileMode modeY = SkTileMode::kDecal;
-  if (p.mode == VGG::TILE_VERTICAL)
-  {
-    modeY = p.mirror ? SkTileMode::kMirror : SkTileMode::kRepeat;
-  }
-  if (p.mode == VGG::TILE_HORIZONTAL)
-  {
-    modeX = p.mirror ? SkTileMode::kMirror : SkTileMode::kRepeat;
-  }
-  if (p.mode == VGG::TILE_BOTH)
-  {
-    modeY = p.mirror ? SkTileMode::kMirror : SkTileMode::kRepeat;
-    modeX = p.mirror ? SkTileMode::kMirror : SkTileMode::kRepeat;
-  }
-  auto m = glm::mat3{ 1.0 };
-  m = glm::rotate(m, p.rotation);
-  m = glm::scale(m, { p.scale, p.scale });
-  const auto        mat = toSkMatrix(m);
-  SkSamplingOptions opt(SkFilterMode::kLinear);
-  auto              shader = img->makeShader(modeX, modeY, opt, &mat);
-  if (auto colorFilter = makeColorFilter(p.imageFilter); shader && colorFilter)
-  {
-    return shader->makeWithColorFilter(colorFilter);
-  }
-  return shader;
-}
-
-inline sk_sp<SkShader> makePatternShader(const Bound& bound, const Pattern& pattern)
-{
-  sk_sp<SkShader> shader;
-  std::visit(
-    Overloaded{ [&](const PatternFill& p) { shader = makeFillPattern(bound, p); },
-                [&](const PatternFit& p) { shader = makeFitPattern(bound, p); },
-                [&](const PatternStretch& p) { shader = makeStretchPattern(bound, p); },
-                [&](const PatternTile& p) { shader = makeTilePattern(bound, p); } },
-    pattern.instance);
-  return shader;
-}
-
-inline sk_sp<SkShader> makeGradientLinear(const Bound& bound, const GradientLinear& g)
-{
-  if (g.stops.empty())
-    return nullptr;
-  auto minPosition = g.stops.front().position;
-  auto maxPosition = g.stops.back().position;
-  // clampPairByLimits(minPosition, maxPosition, 0.f, 1.f, 0.0001f);
-
-  auto f = bound.map(bound.size() * g.from);
-  auto t = bound.map(bound.size() * g.to);
-  auto start = glm::mix(f, t, minPosition);
-  auto end = glm::mix(f, t, maxPosition);
-
-  SkPoint pts[2] = {
-    { (SkScalar)start.x, (SkScalar)start.y },
-    { (SkScalar)end.x, (SkScalar)end.y },
-  };
-  std::vector<SkColor>  colors;
-  std::vector<SkScalar> positions;
-  for (auto it = g.stops.begin(); it != g.stops.end(); ++it)
-  {
-    colors.push_back(it->color);
-    auto p = it->position;
-    positions.push_back((p - minPosition) / (maxPosition - minPosition));
-  }
-  SkMatrix mat = SkMatrix::I();
-  auto     s = SkGradientShader::MakeLinear(
-    pts,
-    colors.data(),
-    positions.data(),
-    colors.size(),
-    SkTileMode::kClamp,
-    0,
-    &mat);
-  return s;
-}
-
-template<typename G>
-inline sk_sp<SkShader> makeGradientRadial(const Bound& bound, const G& g)
-{
-  if (g.stops.empty())
-    return nullptr;
-  auto minPosition = g.stops.front().position;
-  auto maxPosition = g.stops.back().position;
-
-  auto     f = bound.map(bound.size() * g.from);
-  auto     t = bound.map(bound.size() * g.to);
-  auto     start = glm::mix(f, t, minPosition);
-  auto     end = glm::mix(f, t, maxPosition);
-  SkScalar r = glm::distance(end, start);
-
-  std::vector<SkColor>  colors;
-  std::vector<SkScalar> positions;
-  for (auto it = g.stops.begin(); it != g.stops.end(); ++it)
-  {
-    colors.push_back(it->color);
-    auto p = it->position;
-    positions.push_back((p - minPosition) / (maxPosition - minPosition));
-  }
-  auto mat = makeMatrix(bound, g.from, g.to, g.ellipse);
-  return SkGradientShader::MakeRadial(
-    { start.x, start.y },
-    r,
-    colors.data(),
-    positions.data(),
-    colors.size(),
-    SkTileMode::kClamp,
-    0,
-    &mat);
-}
-
-template<typename G>
-inline sk_sp<SkShader> makeGradientAngular(const Bound& bound, const G& g)
-{
-  if (g.stops.empty())
-    return nullptr;
-
-  const auto            minPosition = g.stops.front().position;
-  const auto            maxPosition = g.stops.back().position;
-  auto                  minPosColor = g.stops.front().color;
-  auto                  maxPosColor = g.stops.back().color;
-  std::vector<SkColor>  colors;
-  std::vector<SkScalar> positions;
-  size_t                sz = g.stops.size();
-
-  if (minPosition > 0)
-  {
-    auto c = lerp(minPosColor, maxPosColor, (float)minPosition / (minPosition + 1 - maxPosition));
-    colors.push_back(c);
-    positions.push_back(0);
-    sz += 1;
-  }
-  for (auto iter = g.stops.begin(); iter != g.stops.end(); ++iter)
-  {
-    colors.push_back(iter->color);
-    positions.push_back(iter->position);
-  }
-  if (maxPosition < 1)
-  {
-    auto c = lerp(minPosColor, maxPosColor, (float)minPosition / (minPosition + 1 - maxPosition));
-    colors.push_back(c);
-    positions.push_back(1);
-    sz += 1;
-  }
-
-  const auto mat = makeMatrix(bound, g.from, g.to, g.ellipse);
-  const auto from = bound.map(bound.size() * g.from);
-  return SkGradientShader::MakeSweep(from.x, from.y, colors.data(), positions.data(), sz, 0, &mat);
-}
-
-inline sk_sp<SkShader> makeGradientShader(const Bound& bound, const Gradient& gradient)
-{
-  sk_sp<SkShader> shader;
-  std::visit(
-    Overloaded{
-      [&](const GradientLinear& p) { shader = makeGradientLinear(bound, p); },
-      [&](const GradientRadial& p) { shader = makeGradientRadial(bound, p); },
-      [&](const GradientAngular& p) { shader = makeGradientAngular(bound, p); },
-      [&](const GradientDiamond& p) { shader = makeGradientRadial(bound, p); },
-      [&](const GradientBasic& p) { // TODO::
-      },
-    },
-    gradient.instance);
-  return shader;
-}
-
-inline sk_sp<SkImageFilter> makeBlendModeFilter(EBlendMode blendMode)
-{
-  switch (blendMode)
-  {
-    case VGG::BM_BLEND_DIVIDE:
-      SkImageFilters::Arithmetic(0, 0, 0, 0, true, nullptr, nullptr);
-    default:
-      return nullptr;
-  }
-}
-
-inline void populateSkPaint(
-  const FillType&       fillType,
-  const ContextSetting& st,
-  const SkRect&         rect,
-  SkPaint&              paint)
-{
-  Bound bound{ rect.x(), rect.y(), rect.width(), rect.height() };
-  std::visit(
-    Overloaded{ [&](const Gradient& g)
-                {
-                  paint.setShader(makeGradientShader(bound, g));
-                  paint.setAlphaf(st.opacity);
-                },
-                [&](const Color& c)
-                {
-                  paint.setColor(c);
-                  paint.setAlphaf(c.a * st.opacity);
-                },
-                [&](const Pattern& p)
-                {
-                  paint.setShader(makePatternShader(bound, p));
-                  paint.setAlphaf(st.opacity);
-                } },
-    fillType);
-  auto bm = toSkBlendMode(st.blendMode);
-  if (bm)
-  {
-    std::visit(
-      Overloaded{ [&](const sk_sp<SkBlender>& blender) { paint.setBlender(blender); },
-                  [&](const SkBlendMode& mode) { paint.setBlendMode(mode); } },
-      *bm);
-  }
-}
-
-inline void populateSkPaint(const Border& border, const SkRect& bound, SkPaint& paint)
-{
-  paint.setStyle(SkPaint::kStroke_Style);
-  paint.setPathEffect(
-    SkDashPathEffect::Make(border.dashedPattern.data(), border.dashedPattern.size(), 0));
-  paint.setStrokeJoin(toSkPaintJoin(border.lineJoinStyle));
-  paint.setStrokeCap(toSkPaintCap(border.lineCapStyle));
-  paint.setStrokeMiter(border.miterLimit);
-  paint.setStrokeWidth(border.thickness);
-  populateSkPaint(border.type, border.contextSettings, bound, paint);
 }
 
 inline std::ostream& operator<<(std::ostream& os, const SkMatrix& mat)
